[CBRD-25825] Fix windows build: Add dependency to configure JDK before building libcubridcs and libcubrid (#5942)
http://jira.cubrid.org/browse/CBRD-25825

In Windows builds, there have been instances where the sa and cubrid libraries, which require the JVM, are built before the JDK including the JVM is set up, leading to failures. This issue has been identified in the QA system. To address this, the build is restructured to remove dependencies between the sa, cubrid, cs libraries and the JVM, ensuring a successful build.

- Remove ${JVM_LIBS} from cs, sa, and cubrid libraries where JVM-related features are not used.
- Add JAVA-related INCLUDE and LIBRARY dependencies only when generating the cub_pl binary.
- Due to dependency modifications, separate header files, delete header includes, and modify the code.