[CBRD-25940] Fix JVM's ListenerThread could terminate abnormally due to a system error (#5959)
http://jira.cubrid.org/browse/CBRD-25940

In the ListenerThread, JVM and system-related errors such as OutOfMemory can occur, which may result in the process terminating without collecting Java logs due to an IOException. In such cases, the JVM might not be able to receive any requests from the DB server and could end up as a non-functional empty process.

This PR modifies the behavior to allow the system to terminate itself and restart the DB server if an abnormal error occurs. It enables sending a sigabort signal and restarting through Java's Process.

- Send a signal to its own process if Throwable occurs repeatedly (after waiting for backoff_times and multiple attempts).
- Server.flushLog() is added to forcibly write the termination log and allow the process to terminate.
- Add handling for SIGTERM in pl.cpp