[CBRD-25909] (Spec Change) Transfer permissions granted by the previous owner to the new owner when changing the owner of a Class/Vclass or Stored Procedure. (#5899)
http://jira.cubrid.org/browse/CBRD-25909

In the CBRD-25627 issue, all privileges of the previous owner were revoked, but this may not have been the intended action by the user. Therefore, the existing privileges are now transferred to the new owner as they are

Considerations for Privilege Transfer

If the new owner already possesses the privileges granted by the previous owner, those privileges will be removed during the ownership transfer.
(In CUBRID, a user cannot directly grant or revoke privileges to/from themselves. This measure prevents issues where the new owner would end up with privileges that they are not allowed to have after the transfer.)