[CBRD-25457] backport: Executing the “AUTO_INCREMENT” and “DEFAULT” attributes separately using the “ALTER TABLE MODIFY” syntax on the same column resulted in an error of two attributes co-existing in that column (#5364) (#5928)
http://jira.cubrid.org/browse/CBRD-25457

The following error occurs when the 'AUTO_INCREMENT' and 'DEFAULT' properties are used at the same time when executing the 'CREATE TABLE' and 'ALTER COLUMN' statements.
ERROR: SHARED, DEFAULT and AUTO_INCREMENT cannot be defined with each other.

However, when using the 'ALTER TABLE .. MODIFY' statement to execute 'AUTO_INCREMENT' and 'DEFAULT' attributes separately, the two attributes can coexist on the same column.

Therefore, we modified it so that only the last-applied 'AUTO_INCREMENT' and 'DEFAULT' attribute is applied.