[CBRD-25939] Fix the issue where a synonym retains a duplicate name when the owner of the target class/vclass is changed. (#5935)
http://jira.cubrid.org/browse/CBRD-25939

When creating a Synonym, you cannot use the same unique name as an existing object (Class or Vclass).

However, there was an issue where changing the owner of a synonym's target class to match the synonym caused a duplicate unique name to be created.
This issue has now been fixed.