[CBRD-25968] Put in request cubmethod::header into bootstrap_request struct (#6029)
http://jira.cubrid.org/browse/CBRD-25968

There is no need to create a header every time a bootstrap_request is sent to the PL server; instead, it should be included in the bootstrap_request struct. Additionally, a Valgrind report indicated that std::move(header) might have caused uninitialized bytes due to the object's lifetime not being guaranteed. This could be a memory wrapper issue, but since the reason is unclear, defensive code will be added.

- Include cubmethod::header as a member variable in bootstrap_request.
- Implement the constructor and pack function.