[CBRD-25947] Problem with changing rownum to orderby_num even though order_by is deleted during view merge (#6038)
In the process of converting order_by during view merge, rownum and orderby_num are changed unnecessarily. Conversion is performed only for the following cases.

case 1: order by of mainquery is newly added ==> rownum of mainquery is changed to orderby_num
case 2: order by of subquery is totally removed ==> orderby_num of subquery is changed to rownum