[CBRD-25983] Modify native lib loading for JNI to be allowed only for SP that use a statically loaded class loader (#6057)
http://jira.cubrid.org/browse/CBRD-25983

In the class loader that dynamically loads classes, if a native library for JNI is loaded, a duplicate native library loading may occur on subsequent calls, requiring a restart of the PL server. Therefore, the class loader for dynamic loading should be modified to prevent it from loading native libraries.

- Modify to throw an SecurityException in SpSecurityManager.checkLink() when dynamic class loader try to load native library.