[CBRD-26003] Fix assert error caused by scan manager's open() does not expect error code (#6092)
http://jira.cubrid.org/browse/CBRD-26003

In the scan manager, the open() function does not return an error code directly but triggers an assertion failure on error. Instead, the scan manager relies on the error set by the error manager and determines whether an error occurred for a SCAN_ID by calling the er_errid() function.