[CBRD-25447] Support parallel heap scan (#6141)
Purpose
To improve the performance of heap scans, we plan to add a parallel heap scan feature to CUBRID.

Implementation
This is a merge pull request for the feature branch.

Remarks
Since this is a merge pull request for the feature branch, the code has already been reviewed.
It includes the following pull requests.
#5850
#6093
#6091
#6154
#6166
#6155