[CBRD-26087] [Regression] Core dumped in qfile_reopen_list_as_append_mode at src/query/list_file.c:1348 (#6204)
Purpose

A core dump occurs in debug mode due to assert(0) when the list-id is not initialized inside the reopen() function. This change addresses the issue.

Implementation

Modified the logic to prevent reopen from being called when list_id is uninitialized or when the operation is not based on the mergable-list method.

Remarks

N/A