[CBRD-26088] [Regression][shell] Core dumped in file_tempcache_pop_tran_file at src/storage/file_manager.c:9521 (#6212)
In the original CUBRID source, during the execution of a holdable query, if the value of qentry_p->list_id->page_cnt was 0, the query was considered to have not created a temp file, and thus the corresponding temp file was not preserved.

However, with the introduction of parallel heap scan, a temp file containing results may exist even if it is not directly linked to the list_id. Such files also need to be managed along with the corresponding qentry.

Accordingly, the logic that skipped preserving the temp file when qentry_p->list_id->page_cnt was 0 has been removed.