[CBRD-26099] Join method hint ignored for LEFT OUTER JOIN with empty inline view result (#6237)
http://jira.cubrid.org/browse/CBRD-26099

Purpose :
Fix an issue where join method hints (USE_HASH, USE_MERGE) were not respected when performing a LEFT OUTER JOIN with an inline view that produces zero rows.

Implementation :
Previously, in cases with transitive join conditions, the optimizer always selected a Nested-Loop join regardless of user-provided hints, assuming it was the most appropriate. This change modifies that behavior to allow the optimizer to honor the join method explicitly specified by the user.