[CBRD-25887] Problem with NULL_LOCK being output in lock timeout log (#5873)
http://jira.cubrid.org/browse/CBRD-25887

There is a problem that NULL_LOCK is printed in the lock timeout log. The reason is that the grant_mode of the waiter of the lock resources is printed instead of blocked_mode. In the case of the waiter, it is correct to print the required lock mode, blocked_mode. Also, since the lock is granted in the linked list order, the locks in front are printed as 'blocker', and the locks in the back are printed as 'other waiters'.

In addition, fix the problem that SQL is not printed when a table lock is created by lock escalation.