[CBRD-26100] Enhance parallel heap scan to support row-by-row execution when the select list includes stored procedures (#6222)
http://jira.cubrid.org/browse/CBRD-26100

Purpose

In the current implementation (Parallel Heap Scan), when the SELECT list includes PLCSQL or JavaSP, each thread fetches rows and sends them to the main thread using the gather:row-by-row method. Then, in the end-one-iteration phase, the main thread invokes the PLCSQL or JavaSP based on the received row.

However, due to concerns about malfunctions and test case (TC) instability, this functionality is currently disabled.

We plan to lift this restriction and proceed with stabilizing the related test cases.