[CBRD-25531] When an analytic function is included in a non-mergeable inline view, unreferenced select list items are not removed (#5426)
http://jira.cubrid.org/browse/CBRD-25531

In the rewriter, when a query cannot be merged into a view, the select list of an inline view that is not referenced by the main query can be removed. Or, if the select list of the inline view is not referenced at all by the main query, it can be replaced with a PT_VALUE node with a value of 1.

However, if the inline view contains an analytic function, neither of these actions is performed, which results in either unnecessary columns being retrieved or inefficient execution of the analytic function.