[CBRD-26137] [Regression] Core dumped in pgbuf_is_valid_page at src/storage/page_buffer.c:10186 (#6290)
http://jira.cubrid.org/browse/CBRD-26137

Purpose

In add_list_id, after fixing the last page of head_list_id (which, in the current context, refers to xasl->list_id operating according to legacy behavior), the first vpid of the newly added list_id is linked to the next of that page.

However, the list_id objects created by parallel threads have the NOT_USE_MEMBUF flag set, meaning their vpid does not point to a page in the memory buffer (membuf).

In contrast, the legacy xasl->list_id uses the memory buffer, so its vpid has a volume number set to -1.

To address this, when attempting to fix a vpid in this state, the logic is modified to fix the page in the memory buffer accordingly.