[CBRD-26206] Disable uncorrelated subquery parallel execution when SP exists in preds (#6340)
http://jira.cubrid.org/browse/CBRD-26206

Purpose

Currently, during uncorrelated subquery parallel execution, execution is abandoned if the select-list contains JAVASP/PLCSQL. However, if JAVASP/PLCSQL is included in the predicate, execution is still attempted, which may lead to a hang. To address this issue, the existing logic that only checked the select-list has been modified to iterate through all predicates in the XASL to detect the presence of stored procedures.