[CBRD-26151] Modify the db_auth catalog view to output only the privileges that correspond to the grantor–grantee relationship of the logged-in user. (#6387)
http://jira.cubrid.org/browse/CBRD-26151

During test case (TC) modification, the following issue was identified. (CBRD-25923)

In CUBRID versions 9.3 ~ 11.4, the db_auth view has a problem where if a logged-in user holds only the SELECT privilege on a table, all privileges on that table are displayed.
The same issue occurs for stored procedures, where all privileges are shown if the user only has the EXECUTE privilege.

To resolve this, the db_auth view was modified to restrict the scope so that only the privileges in which the logged-in user (CURRENT_USER) is involved as either the grantor or the grantee are returned.

After the fix, behavior is classified into the following four cases:
1. DBA or DBA group member
- Can check privileges on all objects.
2. OWNER or OWNER group member
- Can check all privileges on objects owned by the user.
- OWNER group members can check all privileges on the OWNER’s objects.
3. Grantor
- Can check privileges granted by the logged-in user.
4. Grantee
- Can check privileges granted to the logged-in user.