[CBRD-26370] Optimize COUNT queries during parallel heap scan (#6606)
Purpose

We implemented optimizations for COUNT(*) and COUNT(column) aggregate operations in parallel heap scans.
Previously, COUNT queries were processed in a row-by-row manner, but now they are executed in parallel to improve performance.
In particular, we enhanced the system so that the COUNT DISTINCT operation used during UPDATE STATISTICS can also be processed in parallel.